feat: Extract test data
